requirements.txt
discord.py>=2.3.2
motor>=3.3.2
python-dotenv>=1.0.0
pymongo>=4.6.1
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
stripe>=8.0.0
aiohttp>=3.9.1
pytz>=2024.1
